+++ R/count.R
@@ -550,20 +550,12 @@ setMethod("hist", "count", function(x, ...) hist(x@stat[,"count"], ...) )
 
 #' @rdname count-method
 setMethod("count", "remote_corpus", function(.Object, ...){
-  ocpu_exec(
-    fn = "count",
-    server = .Object@server,
-    user = .Object@user,
-    password = .Object@password,
-    do.call = FALSE,
-    .Object = .Object@corpus,
-    ...
-  )
+  ocpu_exec(fn = "count", corpus = .Object@corpus, server = .Object@server, restricted = .Object@restricted, .Object = as(.Object, "corpus"), ...)
 })
 
 
 #' @rdname count-method
 setMethod("count", "remote_subcorpus", function(.Object, ...){
-  ocpu_exec(fn = "count", server = .Object@server, do.call = FALSE, .Object = as(.Object, "subcorpus"), ...)
+  ocpu_exec(fn = "count", corpus = .Object@corpus, server = .Object@server, restricted = .Object@restricted, .Object = as(.Object, "subcorpus"), ...)
 })