+++ R/opencpu.R
@@ -8,11 +8,13 @@
 #' @param fn Name of the function/method to execute on remote server (length-one
 #'   \code{character} vector).
 #' @param server The IP/URL of the remote OpenCPU server.
-#' @param user A username required for authentication, if necessary.
-#' @param password A password required for authentication, if necessary.
+#' @param corpus A length-one \code{character} vector, the id of the corpus to be 
+#'   queried.
+#' @param restricted A \code{logical} value, whether credentials are required to
+#'   access the data.
 #' @param do.call Logical, if \code{TRUE}, the function \code{fn} is passed into
 #'   a call of \code{do.call}, which offers some flexibility.
-#' @param ... Further arguments passed into the method/function call.
+#' @param ... Arguments passed into the method/function call.
 #' @export ocpu_exec
 #' @rdname ocpu_exec
 #' @aliases opencpu
@@ -25,45 +27,46 @@
 #'   do.call = TRUE,
 #'   pkg = "polmineR"
 #' )
-#' 
-#' ocpu_exec(
-#'   fn = "subset",
-#'   server = Sys.getenv("OPENCPU_SERVER"),
-#'   do.call = TRUE,
-#'   x = iris,
-#'   subset = substitute(Sepal.Width > 4.0)
-#' )
 #' }
-ocpu_exec <- function(fn, server, user = NULL, password = NULL, do.call = FALSE, ...){
+ocpu_exec <- function(fn, corpus, server, restricted = FALSE, do.call = FALSE, ...){
   if (!requireNamespace("httr", quietly = TRUE))
-    stop("To access a remote corpus, package 'httr' is required, but it is not yet installed.")
+    stop("To access a remote corpus, package 'httr' is required. The 'httr' package is not installed.")
   if (!requireNamespace("curl", quietly = TRUE))
     stop("To access a remote corpus, package 'curl' is required, but it is not yet installed.")
   if (!requireNamespace("protolite", quietly = TRUE))
     stop("To access a remote corpus, package 'protolite' is required, but it is not yet installed.")
 
-  if (!do.call){
-    url <- sprintf("%s/ocpu/library/polmineR/R/%s/pb", server, fn)
-    args <- list(...)
+  url <- if (isFALSE(do.call)){
+    sprintf("%s/ocpu/library/polmineR/R/%s/pb", server, fn)
   } else {
-    url <- sprintf("%s/ocpu/library/base/R/do.call/pb", server)
-    args <- list(what = fn, args = list(...))
+    sprintf("%s/ocpu/library/base/R/do.call/pb", server)
   }
   
   body <- lapply(
     list(...),
     function(x)
-      if (class(x) == "call")
+      if (class(x) == "call"){
         deparse(x)
-    else
-      curl::form_data(protolite::serialize_pb(x), "application/protobuf")
+      } else {
+        curl::form_data(protolite::serialize_pb(x), "application/protobuf")
+      }
   )
-  if (length(user) == 0L && length(password) == 0L){
-    resp <- httr::POST(url = url, body = body)
+  if (isTRUE(restricted)){
+    opencpu_registry <- Sys.getenv("OPENCPU_REGISTRY")
+    if (identical(nchar(opencpu_registry), 0L)){
+      stop("Access to corpora with restricted corpora requires that the environment variable 'OPENCPU_REGISTRY' is set.")
+    }
+    
+    properties <- registry_get_properties(corpus = corpus, registry = opencpu_registry)
+    resp <- httr::POST(
+      url = url, body = body,
+      httr::authenticate(user = properties[["user"]], password = properties[["password"]])
+    )
+    rm(properties)
   } else {
-    resp <- httr::POST(url = url, body = body, httr::authenticate(user = user, password = password))
+    resp <- httr::POST(url = url, body = body)
   }
   httr::stop_for_status(resp)
-  y <- protolite::unserialize_pb(resp$content)
+  protolite::unserialize_pb(resp$content)
 }